+++ packages/args_annotations/CHANGELOG.md
@@ -1,3 +1,35 @@
+## 1.0.2
+
+- Added new static method `builder` to the generated class for creating the associated widget from arguments.
+
 ## 0.0.9
+
+- Added the `args_annotations' package to fix a bug with `source_gen` that does not work with `Flutter` projects.
+
+## 0.0.8
+
+- Parsing fixed for `null` values.
+
+## 0.0.7
+
+- Skip the complex types.
+
+## 0.0.6
+
+- Fix dart:mirrors issue.
+
+## 0.0.5
+
+- Added `super` fields to args.
+
+## 0.0.4
+
+- `SharedPartBuilder` was replaced with `PartBuilder`.
+
+## 0.0.3
+
+- Updated docs.
+
+## 0.0.2
 
 - Initial version.